fixing some bugs

+++ src/Lesko/CodeAnalysis/Binding/Conversion.cs
@@ -46,7 +46,7 @@ namespace Lesko.CodeAnalysis.Binding
                 return Conversion.Implicit;
             }
                 
-            if (from == TypeSymbol.Int && to == TypeSymbol.Float)
+            if (from == TypeSymbol.Float && to == TypeSymbol.Int)
             {
                 return Conversion.Implicit;
             }
